requirements.txt
Flask==3.1.0
line-bot-sdk==3.14.2
gspread==6.1.4
pandas==2.2.3
pymongo==4.10.1
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.163.0
gunicorn